+++ backend/src/price-is-nice/models/guessInformation.ts
